--- a/ce/tests/test_util.py
+++ b/ce/tests/test_util.py
@@ -9,7 +9,10 @@ from numpy.ma import MaskedArray
 from dateutil.parser import parse
 from netCDF4 import Dataset
 
-from ce.api.util import get_array, mean_datetime, open_nc, check_final_cell_method
+from ce.api.util import get_array, mean_datetime, open_nc, check_final_cell_method, get_units_from_run_object
+
+from modelmeta.v2 import Run
+
 
 
 @pytest.fixture(
@@ -94,6 +97,19 @@ def test_open_nc_exception(bad_path):
         with open_nc(bad_path) as nc:
             # Test won't make it this far, but in case we do, let's fail the test
             assert False
+
+@pytest.mark.parametrize(
+    ("run_name", "var_name", "ensemble_name", "expected_units"),
+    [
+        ("run1", "tasmax", "ce", "degC"),
+        ("r1i1p1", "tasmin", "ce", "degC"),
+        ("r1i1p1", "pr", "ce", "kg d-1 m-2"),
+    ],
+)
+def test_get_units_from_run_object(populateddb, run_name, var_name, ensemble_name, expected_units):
+    run_object = Run(name=run_name)
+    units = get_units_from_run_object(populateddb.session, Run, var_name, ensemble_name)
+    assert(units == expected_units)
 
 
 @pytest.mark.parametrize(