--- a/requirements.txt
+++ b/requirements.txt
@@ -3,7 +3,7 @@ Flask-SQLAlchemy
 Flask-Cors
 modelmeta==1.0.0
 shapely>=1.6
-numpy
+numpy~=1.20
 netCDF4==1.5.*
 GDAL~=3.0
 rasterio~=1.1