This PR speeds up the `/data` query by  streamlining `get_units_from_run_object()`.

`get_units_from_run_object()` checks all datasets associated with a particular run and variable to make sure they all have the same units and can be aggregated. Previously, the units were checked incrementally by getting the units from each datafile and comparing them. The rewritten function instead gets a list of all units at once from the database.

Timing information:
| Branch      | time to complete sample /data call |
|-------------|------------------------------------|
| [3.1.0](http://docker-dev01.pcic.uvic.ca:30399/api/data?ensemble_name=ce_files&model=CNRM-CM5&variable=fdd&emission=historical,rcp85&timescale=yearly&time=0&area=)       | 1.372 seconds                      |
| [3.2.0](http://docker-dev01.pcic.uvic.ca:30400/api/data?ensemble_name=ce_files&model=CNRM-CM5&variable=fdd&emission=historical,rcp85&timescale=yearly&time=0&area=)       | 22.135 seconds                     |
| [3.3.0](http://docker-dev01.pcic.uvic.ca:30401/api/data?ensemble_name=ce_files&model=CNRM-CM5&variable=fdd&emission=historical,rcp85&timescale=yearly&time=0&area=)       | 21.860 seconds                     |
| [this branch](http://docker-dev01.pcic.uvic.ca:30402/api/data?ensemble_name=ce_files&model=CNRM-CM5&variable=fdd&emission=historical,rcp85&timescale=yearly&time=0&area=) | 0.509 seconds                      |